seperating migrations and moving models to internal
